Refactor XML model components: consolidate DataModel definition and remove unused files
